requirements.txt
PyPDF2
discord.py
python-dotenv